Moved the functional tests to the tests subfolder
